// src/http/models/modelsVehicleBrand.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API RMI
 * API para gerenciamento de dados de cidadãos do Rio de Janeiro, incluindo autodeclaração de informações e verificação de contato.
 * OpenAPI spec version: 1.0
 */

export interface ModelsVehicleBrand {
  deleted_at?: string
  id?: string
  is_other?: boolean
  name?: string
}