// src/http/models/modelsVehicleInvitationItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API RMI
 * API para gerenciamento de dados de cidadãos do Rio de Janeiro, incluindo autodeclaração de informações e verificação de contato.
 * OpenAPI spec version: 1.0
 */
import type { ModelsConductorStatus } from './modelsConductorStatus'
import type { ModelsVehicleInvitationSummary } from './modelsVehicleInvitationSummary'

export interface ModelsVehicleInvitationItem {
  created_at?: string
  id?: string
  invited_by_cpf?: string
  owner_name?: string
  status?: ModelsConductorStatus
  vehicle?: ModelsVehicleInvitationSummary
  vehicle_id?: string
}