// src/http/models/modelsVehicleRole.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API RMI
 * API para gerenciamento de dados de cidadãos do Rio de Janeiro, incluindo autodeclaração de informações e verificação de contato.
 * OpenAPI spec version: 1.0
 */

export type ModelsVehicleRole =
  (typeof ModelsVehicleRole)[keyof typeof ModelsVehicleRole]

export const ModelsVehicleRole = {
  VehicleRoleOwner: 'owner',
  VehicleRoleConductor: 'conductor',
} as const