// src/http/models/modelsVehicleType.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * API RMI
 * API para gerenciamento de dados de cidadãos do Rio de Janeiro, incluindo autodeclaração de informações e verificação de contato.
 * OpenAPI spec version: 1.0
 */

export type ModelsVehicleType =
  (typeof ModelsVehicleType)[keyof typeof ModelsVehicleType]

export const ModelsVehicleType = {
  VehicleTypeBicicletaEletrica: 'bicicleta_eletrica',
  VehicleTypeAutopropelido: 'autopropelido',
  VehicleTypeCiclomotor: 'ciclomotor',
} as const